+++ src/main/java/net/levelz/access/MobEntityAccess.java
@@ -1,0 +1,9 @@
+package net.levelz.access;
+
+public interface MobEntityAccess {
+
+    public void setSpawnerMob(boolean spawnerMob);
+
+    public boolean isSpawnerMob();
+
+}